Merge remote-tracking branch 'orig-master/master' into HEAD

+++ tools/idf_tools.py
@@ -1185,6 +1185,11 @@ def action_install(args):
 def action_install_python_env(args):
     idf_python_env_path, _, virtualenv_python = get_python_env_path()
 
+    is_virtualenv = hasattr(sys, 'real_prefix') or (hasattr(sys, 'base_prefix') and sys.base_prefix != sys.prefix)
+    if is_virtualenv and (not os.path.exists(idf_python_env_path) or args.reinstall):
+        fatal('This script was called from a virtual environment, can not create a virtual environment again')
+        raise SystemExit(1)
+
     if args.reinstall and os.path.exists(idf_python_env_path):
         warn('Removing the existing Python environment in {}'.format(idf_python_env_path))
         shutil.rmtree(idf_python_env_path)
@@ -1199,7 +1204,7 @@ def action_install_python_env(args):
             subprocess.check_call([sys.executable, '-m', 'pip', 'install', '--user', 'virtualenv'],
                                   stdout=sys.stdout, stderr=sys.stderr)
 
-        subprocess.check_call([sys.executable, '-m', 'virtualenv', '--no-site-packages', idf_python_env_path],
+        subprocess.check_call([sys.executable, '-m', 'virtualenv', idf_python_env_path],
                               stdout=sys.stdout, stderr=sys.stderr)
     run_args = [virtualenv_python, '-m', 'pip', 'install', '--no-warn-script-location']
     requirements_txt = os.path.join(global_idf_path, 'requirements.txt')
